Added Sponsors Page
I've added the sponsors page for us. At the moment, in order to achieve this, I've setup a specific template base for posts and changed the default to a general layout to work for this.

This method doesn't allow translation since functions can't be used in single page content with Hugo.

If someone has a better idea in implementing this, I'm all ears. Feel free to make a PR.